fixed book-image create issue
